One version for all notebooks
I made a lot of trials on all notebooks of this Guided Capstone. Making a lot of errors I finally messed up all of the initial notebooks and was not able to reproduce my results from the last notebook step 5 properly. I also had started the project w/o checking that "Big Mountain Resort" was included in the starter dataset. Therefore I decided to start again and put all content in one file, trying to avoid / correct a lot of the errors I made in the former versions. I still don't understand a lot of the processes (the k-means clustering serves as a feature creation tool, but I don't understand it's deeper meaning for explaining the final value. I also can't tell a story about how to predict the price of the "AdultWeekend" ticket. But I learned a lot...